refactor(RepositoryCache): implement KeyGenerator behavior directly
