+++ lib/mbta_v3_api/repository_cache.ex
@@ -5,5 +5,20 @@ defmodule MBTAV3API.RepositoryCache do
   use Nebulex.Cache,
     otp_app: :mobile_app_backend,
     adapter: Nebulex.Adapters.Local,
-    default_key_generator: MBTAV3API.RepositoryCache.KeyGenerator
+    default_key_generator: __MODULE__
+
+  @behaviour Nebulex.Caching.KeyGenerator
+
+  @impl Nebulex.Caching.KeyGenerator
+  def generate(mod, fun, []) do
+    "#{mod}|#{fun}"
+  end
+
+  def generate(mod, fun, [arg]) do
+    "#{mod}|#{fun}|#{:erlang.phash2(arg)}"
+  end
+
+  def generate(mod, fun, args) do
+    "#{mod}|#{fun}|#{:erlang.phash2(args)}"
+  end
 end
